requirements.txt
numpy==1.24.1
nptyping==2.4.1
pygame==2.1.2
pytest~=7.2.0
PyYAML==6.0
zarr==2.13.3
tqdm==4.64.1
fire==0.5.0
rich==13.0.0
Pillow==9.3.0
opencv-python==4.7.0.68